refactor(toolchain/eslint-config): remove orig eslint-config, replace with new eslint-config
This is the last step to move to the revamped
eslint configuration.
