+++ apps/sveltekit-example-app/package.json
@@ -25,7 +25,7 @@
     "@sveltejs/kit": "2.19.0",
     "@sveltejs/vite-plugin-svelte": "5.0.3",
     "@tailwindcss/vite": "4.0.12",
-    "@toolchain/eslint-config-new": "workspace:*",
+    "@toolchain/eslint-config": "workspace:*",
     "@toolchain/typescript-config": "workspace:*",
     "@toolchain/vitest-config": "workspace:*",
     "http-status-codes": "2.3.0",
