+++ packages/api/package.json
@@ -24,7 +24,7 @@
     "zod": "3.24.2"
   },
   "devDependencies": {
-    "@toolchain/eslint-config-new": "workspace:*",
+    "@toolchain/eslint-config": "workspace:*",
     "@toolchain/typescript-config": "workspace:*",
     "@toolchain/vitest-config": "workspace:*"
   }
