+++ packages/db-drizzlepg/package.json
@@ -30,7 +30,7 @@
   },
   "devDependencies": {
     "@softwaretechnik/dbml-renderer": "1.0.30",
-    "@toolchain/eslint-config-new": "workspace:*",
+    "@toolchain/eslint-config": "workspace:*",
     "@toolchain/typescript-config": "workspace:*",
     "@toolchain/vitest-config": "workspace:*",
     "@types/pg": "8.11.11",
