--- toolchain/eslint-config-new/configs/base.js
@@ -1,24 +1,0 @@
-import { config } from 'typescript-eslint'
-
-import eslintCommentsConfig from './js/eslint-comments.js'
-
-export default config(
-  {
-    name: 'eslint-config:config:base',
-
-    languageOptions: {
-      ecmaVersion: 2025,
-      parserOptions: {
-        ecmaVersion: 2025,
-        sourceType: 'module',
-      },
-      sourceType: 'module',
-    },
-    linterOptions: {
-      noInlineConfig: false,
-      reportUnusedDisableDirectives: 'error',
-      reportUnusedInlineConfigs: 'error',
-    },
-  },
-  eslintCommentsConfig,
-)
