--- toolchain/eslint-config-new/configs/global-ignores.js
@@ -1,9 +1,0 @@
-import { config } from 'typescript-eslint'
-
-export default config([
-  {
-    name: 'eslint-config:config:global-ignores',
-
-    ignores: ['coverage', '.svelte-kit', 'build'],
-  },
-])
