--- toolchain/eslint-config-new/eslint.config.js
@@ -1,1 +1,0 @@
-export { default } from './profile/node.js'
