--- toolchain/eslint-config-new/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "@toolchain/eslint-config-new",
-  "version": "0.0.0",
-  "private": true,
-  "license": "UNLICENSED",
-  "type": "module",
-  "exports": {
-    "./profile/node": "./profile/node.js",
-    "./profile/svelte": "./profile/svelte-profile.js"
-  },
-  "scripts": {
-    "clean": "del .turbo",
-    "lint": "eslint ."
-  },
-  "dependencies": {
-    "@eslint/js": "9.22.0",
-    "@vitest/eslint-plugin": "1.1.36",
-    "eslint": "9.22.0",
-    "eslint-config-prettier": "10.1.1",
-    "eslint-import-resolver-typescript": "3.8.3",
-    "eslint-plugin-eslint-comments": "3.2.0",
-    "eslint-plugin-import": "2.31.0",
-    "eslint-plugin-perfectionist": "4.10.1",
-    "eslint-plugin-prefer-arrow": "1.2.3",
-    "eslint-plugin-promise": "7.2.1",
-    "eslint-plugin-svelte": "3.1.0",
-    "eslint-plugin-unicorn": "57.0.0",
-    "typescript-eslint": "8.26.0"
-  },
-  "devDependencies": {
-    "globals": "16.0.0"
-  }
-}
