+++ toolchain/eslint-config/configs/base.js
@@ -1,43 +1,24 @@
-import { defineFlatConfig } from 'eslint-define-config'
-import eslintCommentsPlugin from 'eslint-plugin-eslint-comments'
-import globals from 'globals'
+import { config } from 'typescript-eslint'
 
-const { es2021, nodeBuiltin } = globals
-export default defineFlatConfig([
+import eslintCommentsConfig from './js/eslint-comments.js'
+
+export default config(
   {
-    linterOptions: {
-      noInlineConfig: false,
-      reportUnusedDisableDirectives: true,
-    },
+    name: 'eslint-config:config:base',
 
     languageOptions: {
-      // TODO: Should this be moved to each profile
-      globals: {
-        ...es2021,
-        ...nodeBuiltin,
+      ecmaVersion: 2025,
+      parserOptions: {
+        ecmaVersion: 2025,
+        sourceType: 'module',
       },
+      sourceType: 'module',
+    },
+    linterOptions: {
+      noInlineConfig: false,
+      reportUnusedDisableDirectives: 'error',
+      reportUnusedInlineConfigs: 'error',
     },
   },
-
-  /**
-   * eslint-plugin-eslint-comments configuration
-   *
-   * Additional ESLint rules for directive comments of ESLint.
-   * https://github.com/mysticatea/eslint-plugin-eslint-comments
-   */
-  {
-    plugins: {
-      'eslint-comments': eslintCommentsPlugin,
-    },
-    rules: {
-      ...eslintCommentsPlugin.configs.recommended.rules,
-
-      /**
-       * Require comments on ESLint disable directives.
-       *
-       * 🚫 Not fixable - https://mysticatea.github.io/eslint-plugin-eslint-comments/rules/require-description.html
-       */
-      'eslint-comments/require-description': 'error',
-    },
-  },
-])
+  eslintCommentsConfig,
+)
