+++ toolchain/eslint-config/configs/ts/import.js
@@ -1,37 +1,35 @@
-import { defineFlatConfig } from 'eslint-define-config'
+import plugin from 'eslint-plugin-import'
+import { config } from 'typescript-eslint'
 
-export default defineFlatConfig({
-  rules: {
-    // TypeScript already provides the same checks as part of standard type checking
-    // see https://typescript-eslint.io/linting/troubleshooting/performance-troubleshooting/#eslint-plugin-import
-    'import/default': 'off',
-    'import/named': 'off',
-    'import/namespace': 'off',
-    'import/no-named-as-default-member': 'off',
+export default config(
+  //
+  plugin.flatConfigs.typescript,
+  {
+    name: 'eslint-config:config:ts:import',
 
-    // TODO: Might want to consider turning these on at least during CI
-    'import/no-cycle': 'off',
-    'import/no-deprecated': 'off',
-    'import/no-named-as-default': 'off',
-    'import/no-unused-modules': 'off',
+    rules: {
+      // TypeScript already provides the same checks as part of standard type checking
+      // see https://typescript-eslint.io/troubleshooting/typed-linting/performance#eslint-plugin-import
+      'import/default': 'off',
+      'import/named': 'off',
+      'import/namespace': 'off',
+      'import/no-named-as-default-member': 'off',
+      'import/no-unresolved': 'off',
 
-    // TypeScript will automatically enforce that you include extensions
-    // when moduleResolution is set to Node16 or NodeNext.
-    'import/extensions': 'off',
-  },
-  settings: {
-    'import/extensions': ['.ts', '.cts', '.mts', '.tsx', '.js', '.cjs', '.mjs', '.jsx'],
-    'import/external-module-folders': ['node_modules', 'node_modules/@types'],
-    'import/parsers': {
-      '@typescript-eslint/parser': ['.ts', '.cts', '.mts', '.tsx'],
+      // TODO: Might want to consider turning these on at least during CI
+      'import/no-cycle': 'off',
+      'import/no-deprecated': 'off',
+      'import/no-named-as-default': 'off',
+      'import/no-unused-modules': 'off',
+
+      // TypeScript will automatically enforce that you include extensions
+      // when moduleResolution is set to Node16 or NodeNext.
+      'import/extensions': 'off',
     },
-    'import/resolver': {
-      node: {
-        extensions: ['.ts', '.cts', '.mts', '.tsx', '.js', '.cjs', '.mjs', '.jsx'],
-      },
-      typescript: {
-        alwaysTryTypes: true,
+    settings: {
+      'import/resolver': {
+        typescript: true,
       },
     },
   },
-})
+)
