+++ toolchain/eslint-config/configs/typescript.js
@@ -1,25 +1,65 @@
-import parser from '@typescript-eslint/parser'
-import { defineFlatConfig } from 'eslint-define-config'
+import { config, configs } from 'typescript-eslint'
 
-import baseTypeScriptConfig, { TYPESCRIPT_FILES } from './ts/base.js'
-import typescriptImportConfig from './ts/import.js'
+import importConfig from './ts/import.js'
 
-export default defineFlatConfig([
+export default config(
+  //
+  importConfig,
   {
-    files: TYPESCRIPT_FILES,
+    name: 'eslint-config:config:typescript',
+
+    extends: [configs.recommendedTypeChecked, configs.stylisticTypeChecked],
+    files: ['**/*.ts', '**/*.cts', '**/*.mts', '**/*.tsx'],
     languageOptions: {
-      parser: parser,
       parserOptions: {
         projectService: true,
       },
     },
+    rules: {
+      '@typescript-eslint/consistent-indexed-object-style': ['error', 'record'],
+      '@typescript-eslint/consistent-type-exports': 'error',
+      '@typescript-eslint/consistent-type-imports': [
+        'error',
+        {
+          disallowTypeAnnotations: true,
+          fixStyle: 'inline-type-imports',
+          prefer: 'type-imports',
+        },
+      ],
+      '@typescript-eslint/explicit-function-return-type': ['error', { allowExpressions: true }],
+      '@typescript-eslint/explicit-module-boundary-types': 'error',
+      '@typescript-eslint/method-signature-style': ['error', 'property'],
+      '@typescript-eslint/no-deprecated': 'error',
+      '@typescript-eslint/no-import-type-side-effects': 'error',
+      '@typescript-eslint/no-non-null-assertion': 'error',
+      '@typescript-eslint/no-redeclare': 'error',
+      '@typescript-eslint/no-shadow': 'error',
+      '@typescript-eslint/no-unused-vars': [
+        'error',
+        {
+          argsIgnorePattern: '^_',
+          caughtErrorsIgnorePattern: '^_',
+          ignoreRestSiblings: true,
+          varsIgnorePattern: '^_',
+        },
+      ],
+      '@typescript-eslint/no-use-before-define': [
+        'error',
+        {
+          classes: false,
+          functions: false,
+          variables: true,
+        },
+      ],
+      '@typescript-eslint/switch-exhaustiveness-check': [
+        'error',
+        {
+          allowDefaultCaseForExhaustiveSwitch: false,
+          requireDefaultForNonUnion: true,
+        },
+      ],
+
+      '@typescript-eslint/no-require-imports': 'error',
+    },
   },
-  {
-    files: TYPESCRIPT_FILES,
-    ...baseTypeScriptConfig,
-  },
-  {
-    files: TYPESCRIPT_FILES,
-    ...typescriptImportConfig,
-  },
-])
+)
