+++ toolchain/eslint-config/package.json
@@ -5,9 +5,8 @@
   "license": "UNLICENSED",
   "type": "module",
   "exports": {
-    "./profile/node": "./profile/node.js",
-    "./profile/svelte": "./profile/svelte.js",
-    "./configs/+aws-appsync": "./configs/+aws-appsync.js"
+    "./profile/node": "./profile/node-profile.js",
+    "./profile/svelte": "./profile/svelte-profile.js"
   },
   "scripts": {
     "clean": "del .turbo",
@@ -15,8 +14,6 @@
   },
   "dependencies": {
     "@eslint/js": "9.22.0",
-    "@typescript-eslint/eslint-plugin": "8.26.0",
-    "@typescript-eslint/parser": "8.26.0",
     "@vitest/eslint-plugin": "1.1.36",
     "eslint": "9.22.0",
     "eslint-config-prettier": "10.1.1",
@@ -26,13 +23,11 @@
     "eslint-plugin-perfectionist": "4.10.1",
     "eslint-plugin-prefer-arrow": "1.2.3",
     "eslint-plugin-promise": "7.2.1",
-    "eslint-plugin-svelte": "2.46.1",
+    "eslint-plugin-svelte": "3.1.0",
     "eslint-plugin-unicorn": "57.0.0",
-    "globals": "16.0.0",
-    "svelte-eslint-parser": "1.0.0"
+    "typescript-eslint": "8.26.0"
   },
   "devDependencies": {
-    "@types/eslint": "9.6.1",
-    "eslint-define-config": "2.1.0"
+    "globals": "16.0.0"
   }
 }
